+++ src/main/java/com/by/form/AdminMemberForm.java
@@ -6,40 +6,40 @@ import java.util.Calendar;
  * Created by yagamai on 15-12-9.
  */
 public class AdminMemberForm {
-    private String mobile;
-    private Long card;
-    private Calendar beginTime;
-    private Calendar endTime;
+	private String mobile;
+	private Integer card;
+	private Calendar beginTime;
+	private Calendar endTime;
 
-    public String getMobile() {
-        return mobile;
-    }
+	public String getMobile() {
+		return mobile;
+	}
 
-    public void setMobile(String mobile) {
-        this.mobile = mobile;
-    }
+	public void setMobile(String mobile) {
+		this.mobile = mobile;
+	}
 
-    public Long getCard() {
-        return card;
-    }
+	public Integer getCard() {
+		return card;
+	}
 
-    public void setCard(Long card) {
-        this.card = card;
-    }
+	public void setCard(Integer card) {
+		this.card = card;
+	}
 
-    public Calendar getBeginTime() {
-        return beginTime;
-    }
+	public Calendar getBeginTime() {
+		return beginTime;
+	}
 
-    public void setBeginTime(Calendar beginTime) {
-        this.beginTime = beginTime;
-    }
+	public void setBeginTime(Calendar beginTime) {
+		this.beginTime = beginTime;
+	}
 
-    public Calendar getEndTime() {
-        return endTime;
-    }
+	public Calendar getEndTime() {
+		return endTime;
+	}
 
-    public void setEndTime(Calendar endTime) {
-        this.endTime = endTime;
-    }
+	public void setEndTime(Calendar endTime) {
+		this.endTime = endTime;
+	}
 }
